Fix linting
