+++ mountaineer/controller.py
@@ -15,7 +15,6 @@ from typing import (
 )
 
 from inflection import underscore
-from pydantic import BaseModel
 
 from mountaineer.actions import (
     FunctionActionType,
@@ -34,23 +33,6 @@ if TYPE_CHECKING:
     from mountaineer.app import ControllerDefinition
 
 RenderInput = ParamSpec("RenderInput")
-
-
-class BuildMetadata(BaseModel):
-    """
-    Controller metadata that is established during build-time and might be stripped
-    during the final output. For instance, layout.tsx files might not be included
-    in production payloads but this metadata object still stores their hierarchy
-    relative to the controllers.
-
-    """
-
-    # All paths in the metadata should be absolute paths, since they are consistent
-    # with regard to the builder filesystem but might be different when deployed
-    view_path: Path
-
-    # Organized by hierarchy, first index is the outermost layout
-    layout_view_paths: list[Path] = []
 
 
 class ControllerBase(ABC, Generic[RenderInput]):
@@ -92,7 +74,6 @@ class ControllerBase(ABC, Generic[RenderInput]):
         # Set by the path resolution layer
         self.view_base_path: Path | None = None
         self.ssr_path: Path | None = None
-        self.build_metadata: BuildMetadata | None = None
 
         self.resolve_paths()
 
@@ -191,14 +172,13 @@ class ControllerBase(ABC, Generic[RenderInput]):
             # Unable to resolve, no-op
             return False
 
-        script_name = underscore(self.__class__.__name__)
         self.view_base_path = view_base
 
         # We'll update this bool if we can't find any dependencies
         found_dependencies = True
 
         # The SSR path is going to be static
-        ssr_path = view_base / "_ssr" / f"{script_name}.js"
+        ssr_path = view_base / "_ssr" / f"{self.script_name}.js"
         if ssr_path.exists():
             self.ssr_path = ssr_path
             ssr_map_path = ssr_path.with_suffix(".js.map")
@@ -209,7 +189,7 @@ class ControllerBase(ABC, Generic[RenderInput]):
             found_dependencies = False
 
         # Find the md5-converted cache path
-        md5_script_pattern = re_compile(script_name + "-" + "[a-f0-9]{32}" + ".js")
+        md5_script_pattern = re_compile(self.script_name + "-" + "[a-f0-9]{32}" + ".js")
         if (view_base / "_static").exists():
             self.bundled_scripts = [
                 path.name
@@ -224,12 +204,8 @@ class ControllerBase(ABC, Generic[RenderInput]):
         else:
             found_dependencies = False
 
-        # Find the metadata
-        metadata_path = view_base / "_metadata" / f"{script_name}.json"
-        if metadata_path.exists():
-            metadata = BuildMetadata.model_validate_json(metadata_path.read_text())
-            self.build_metadata = metadata
-        else:
-            found_dependencies = False
-
         return found_dependencies
+
+    @property
+    def script_name(self):
+        return underscore(self.__class__.__name__)
